Fixed progress bar.
StreamDownloader

Changed:
-> MainWindow.xaml.cs

StreamDownloaderControls

Changed:
-> FlatWindow.cs
-> Themes/Generic.xaml

StreamDownloaderDownload

Changed:
-> Download/DownloadTask.cs
-> Download/FileDownload.cs
-> Download/JSON/StoredDownload.cs

+++ StreamDownloaderDownload/Download/JSON/StoredDownload.cs
@@ -24,7 +24,7 @@ namespace StreamDownloaderDownload.Download.JSON
     public class DownloadStatus
     {
         [DataMember(Name = "WrittenChunks")]
-        public uint WrittenChunks { get; set; }
+        public ulong WrittenChunks { get; set; }
         [DataMember(Name = "ChunkSize")]
         public uint ChunkSize { get; set; }
         [DataMember(Name = "ContentLength")]
